Consolidate election package functions in libs/backend
