+++ libs/backend/package.json
@@ -49,6 +49,7 @@
     "fs-extra": "11.1.1",
     "js-sha256": "^0.9.0",
     "jszip": "^3.9.1",
+    "lodash.merge": "^4.6.2",
     "micromatch": "^4.0.5",
     "readline": "^1.3.0",
     "stream-chain": "^2.2.5",
@@ -62,6 +63,7 @@
     "@types/debug": "4.1.8",
     "@types/fs-extra": "11.0.1",
     "@types/jest": "^29.5.3",
+    "@types/lodash.merge": "^4.6.9",
     "@types/lodash.set": "^4.3.7",
     "@types/micromatch": "^4.0.2",
     "@types/node": "16.18.23",
