+++ libs/utils/package.json
@@ -41,7 +41,6 @@
     "debug": "4.3.4",
     "fast-text-encoding": "^1.0.2",
     "jszip": "^3.9.1",
-    "lodash.merge": "^4.6.2",
     "lodash.setwith": "^4.3.2",
     "luxon": "^3.0.0",
     "moment": "2.29.1",
@@ -56,7 +55,6 @@
     "@types/fast-text-encoding": "^1.0.1",
     "@types/jest": "^29.5.3",
     "@types/kiosk-browser": "workspace:*",
-    "@types/lodash.merge": "^4.6.9",
     "@types/lodash.setwith": "^4.3.9",
     "@types/luxon": "^3.0.0",
     "@types/node": "16.18.23",
